Front

See merge request s03-webmobile2-sub3/s03p13a108!98

+++ frontend/src/components/feed/item/Writer.vue
@@ -20,12 +20,12 @@
       class="text-decoration-none"
     >
       <v-list-item-content>
-        <v-list-item-title class="text-left red--text text--lighten-2">
-          {{ user.unick }}
-        </v-list-item-title>
-        <v-list-item-subtitle class="text-left red--text text--lighten-2">
-          {{ user.uid }}
-        </v-list-item-subtitle>
+        <v-list-item-title class="text-left red--text text--lighten-2">{{
+          user.unick
+        }}</v-list-item-title>
+        <v-list-item-subtitle class="text-left red--text text--lighten-2">{{
+          user.uid
+        }}</v-list-item-subtitle>
       </v-list-item-content>
     </router-link>
     <v-spacer></v-spacer>
@@ -37,12 +37,15 @@
       </template>
 
       <v-list class="text-center">
-        <v-list-item v-if="user.uid === authUserUid" @click="moveToUpdateFeed">
+        <v-list-item
+          v-if="user.unick === authUserUnick"
+          @click="moveToUpdateFeed"
+        >
           <v-list-item-title class="blue--text text-lighten-2"
             >게시글 수정</v-list-item-title
           >
         </v-list-item>
-        <v-list-item v-if="user.uid === authUserUid" @click="deleteFeedNow">
+        <v-list-item v-if="user.unick === authUserUnick" @click="deleteFeedNow">
           <v-list-item-title class="red--text text-lighten-2"
             >게시글 삭제</v-list-item-title
           >
@@ -63,7 +66,8 @@
 </template>
 
 <script>
-import { mapActions, mapState } from "vuex";
+import swal from "sweetalert";
+import { mapActions, mapGetters } from "vuex";
 import SERVER from "@/api/api";
 
 export default {
@@ -73,7 +77,7 @@ export default {
     item: Boolean,
   },
   computed: {
-    ...mapState("accounts", ["authUserUid"]),
+    ...mapGetters("accounts", ["authUserUnick"]),
     imgName() {
       if (this.user.uprofile) {
         return this.user.uprofile.name;
@@ -95,7 +99,17 @@ export default {
     },
 
     deleteFeedNow() {
-      this.deleteFeed(this.$route.params.fid);
+      swal({
+        title: "삭제하시겠습니까?",
+        text: "이 작업은 취소 할 수 없습니다.",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      }).then((doDelete) => {
+        if (doDelete) {
+          this.deleteFeed(this.$route.params.fid);
+        }
+      });
     },
   },
 };
